python-auth-service/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0
psycopg2-binary==2.9.7
redis==4.6.0
requests==2.31.0
PyMySQL==1.1.0
cryptography==41.0.4
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1 